Move picongpu/simulationControl to picongpu/simulation/control
