+++ include/picongpu/particles/creation/creation.kernel
@@ -25,7 +25,7 @@
 #include <pmacc/mappings/kernel/AreaMapping.hpp>
 #include "picongpu/particles/ParticlesInit.kernel"
 #include <pmacc/mappings/simulation/GridController.hpp>
-#include "picongpu/simulationControl/MovingWindow.hpp"
+#include "picongpu/simulation/control/MovingWindow.hpp"
 #include <pmacc/traits/Resolve.hpp>
 #include <pmacc/math/vector/Int.hpp>
 #include <pmacc/nvidia/atomic.hpp>
