+++ include/picongpu/plugins/hdf5/HDF5Writer.hpp
@@ -56,7 +56,7 @@
 #include <pmacc/mappings/simulation/SubGrid.hpp>
 #include <pmacc/dimensions/GridLayout.hpp>
 #include <pmacc/pluginSystem/PluginConnector.hpp>
-#include "picongpu/simulationControl/MovingWindow.hpp"
+#include "picongpu/simulation/control/MovingWindow.hpp"
 #include <pmacc/math/Vector.hpp>
 
 #include "picongpu/plugins/output/IIOBackend.hpp"
