+++ include/picongpu/unitless/starter.unitless
@@ -21,8 +21,8 @@
 
 #include "picongpu/initialization/InitialiserController.hpp"
 #include "picongpu/plugins/PluginController.hpp"
-#include "picongpu/simulationControl/MySimulation.hpp"
-#include "picongpu/simulationControl/SimulationStarter.hpp"
+#include "picongpu/simulation/control/MySimulation.hpp"
+#include "picongpu/simulation/control/SimulationStarter.hpp"
 
 
 namespace picongpu
