+++ share/picongpu/examples/ThermalTest/include/picongpu/unitless/starter.unitless
@@ -22,7 +22,7 @@
 #include "picongpu/plugins/PluginController.hpp"
 #include "picongpu/initialization/InitialiserController.hpp"
 #include "picongpu/ThermalTestSimulation.hpp"
-#include "picongpu/simulationControl/SimulationStarter.hpp"
+#include "picongpu/simulation/control/SimulationStarter.hpp"
 
 
 namespace picongpu
